Fix validator implementation in certificate controller

+++ module2/gift-certificate-system/service/src/main/java/com/epam/esm/service/GiftCertificateServiceImpl.java
@@ -3,12 +3,13 @@ package com.epam.esm.service;
 import com.epam.esm.dao.GiftCertificateDao;
 import com.epam.esm.dao.TagDao;
 import com.epam.esm.exception.GiftCertificateNotFoundException;
+import com.epam.esm.exception.TagNotFoundException;
 import com.epam.esm.model.GiftCertificate;
 import com.epam.esm.model.Tag;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.time.LocalDateTime;
+import java.text.MessageFormat;
 import java.time.ZonedDateTime;
 import java.util.List;
 import java.util.Optional;
@@ -27,7 +28,12 @@ public class GiftCertificateServiceImpl implements GiftCertificateService {
 
     @Override
     public List<GiftCertificate> findAllCertificates() {
-        return giftCertificateDao.findAll();
+        List<GiftCertificate> certificateList = giftCertificateDao.findAll();
+        if (certificateList.isEmpty()) {
+            throw new GiftCertificateNotFoundException("Gift certificates were not found");
+        } else {
+            return certificateList;
+        }
     }
 
     @Override
@@ -43,7 +49,6 @@ public class GiftCertificateServiceImpl implements GiftCertificateService {
     @Override
     public void updateCertificate(GiftCertificate giftCertificate) {
         List<Tag> tags = giftCertificate.getTags();
-
         tags.forEach(tag -> tagDao.assignNewTagToCertificate(tag.getId(), giftCertificate.getId()));
         giftCertificate.setLastUpdateDate(ZonedDateTime.now());
         giftCertificateDao.update(giftCertificate);
@@ -56,33 +61,64 @@ public class GiftCertificateServiceImpl implements GiftCertificateService {
             GiftCertificate giftCertificate = certificate.get();
             giftCertificate.getTags().forEach(tag -> tagDao.removeTagAndCertificate(tag.getId(), id));
         } else {
-            throw new GiftCertificateNotFoundException("Gift certificate with id: " + id + " not found");
+            throw new GiftCertificateNotFoundException(MessageFormat.format("Gift certificate with id: {0} not found", id));
         }
     }
 
     @Override
     public List<GiftCertificate> getCertificatesByTagName(String tagName) {
-        return giftCertificateDao.getCertificatesByTagName(tagName);
+        List<GiftCertificate> certificateList = giftCertificateDao.getCertificatesByTagName(tagName);
+        if (certificateList.isEmpty()) {
+            throw new GiftCertificateNotFoundException(MessageFormat.format("Gift certificate with tag name: {0} was not found", tagName));
+        } else {
+            return certificateList;
+        }
     }
 
     @Override
     public List<GiftCertificate> getCertificatesByPartOfName(String partOfName) {
-        return giftCertificateDao.getCertificatesByPartOfName(partOfName);
+        List<GiftCertificate> certificateList = giftCertificateDao.getCertificatesByPartOfName(partOfName);
+        if (certificateList.isEmpty()) {
+            throw new GiftCertificateNotFoundException(MessageFormat.format("Gift certificate containing the name: {0} was not found", partOfName));
+        } else {
+            return certificateList;
+        }
     }
 
     @Override
     public List<GiftCertificate> getCertificatesByPartOfDescription(String partOfDescription) {
-        return giftCertificateDao.getCertificatesByPartOfDescription(partOfDescription);
+        List<GiftCertificate> certificateList = giftCertificateDao.getCertificatesByPartOfDescription(partOfDescription);
+        if (certificateList.isEmpty()) {
+            throw new GiftCertificateNotFoundException(MessageFormat.format("Gift certificate containing the description: {0} was not found", partOfDescription));
+        } else {
+            return certificateList;
+        }
     }
 
     @Override
     public void addTagToCertificate(Long certificateId, Long tagId) {
-        tagDao.addNewTagAndToCertificate(tagId, certificateId);
+        Optional<GiftCertificate> certificate = giftCertificateDao.find(certificateId);
+        Optional<Tag> tag = tagDao.find(tagId);
+        if (!certificate.isPresent()) {
+            throw new GiftCertificateNotFoundException(MessageFormat.format("Gift certificate with id: {0} was not found", certificateId));
+        } else if (!tag.isPresent()) {
+            throw new TagNotFoundException(MessageFormat.format("Tag with id: {0} was not found", tagId));
+        } else {
+            tagDao.addNewTagAndToCertificate(tagId, certificateId);
+        }
     }
 
     @Override
     public void removeTagFromCertificate(Long certificateId, Long tagId) {
-        giftCertificateDao.removeTagFromCertificate(certificateId, tagId);
+        Optional<GiftCertificate> certificate = giftCertificateDao.find(certificateId);
+        Optional<Tag> tag = tagDao.find(tagId);
+        if (!certificate.isPresent()) {
+            throw new GiftCertificateNotFoundException(MessageFormat.format("Gift certificate with id: {0} was not found", certificateId));
+        } else if (!tag.isPresent()) {
+            throw new TagNotFoundException(MessageFormat.format("Tag with id: {0} was not found", tagId));
+        } else {
+            giftCertificateDao.removeTagFromCertificate(certificateId, tagId);
+        }
     }
 
     @Override
